+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.4.0a1.post1] - 2022-11-20
 ### Changed
 - [CacheStrategy.__init__][yuyo.list_status.CacheStrategy.__init__] now takes two arguments
   `(hikari.api.Cache, hikari.ShardAware)`.